pyproject.toml
[project]
name = "retro-ai"
version = "0.1.0"
description = "RL training framework for retro game emulators"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=1.0.0",
    "stable-baselines3>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "torch>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=24.0",
    "ruff>=0.4.0",
    "pybind11>=3.0",
]
video = [
    "opencv-python>=4.0",
]

[project.scripts]
retro-ai = "retro_ai.training.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
